requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.2
pydantic>=2.5.0
python-dotenv>=1.0.1
requests>=2.31.0
python-telegram-bot==20.7
typing-extensions>=4.8.0
